chore: move classes to own dirs and use class methods in views -- still has bug related to class methods
