+++ academic_calendar_scraper/course_planning_tool/utils/course/course.py
