+++ academic_calendar_scraper/course_planning_tool/utils/prereqtree/prereqtree.py
@@ -1,4 +1,4 @@
-from course import Course
+from ..course.course import Course
 
 
 class PrereqTree:
